Move CompetitionsViewAdapter to separate file
